Removing custom validator logic in favoro f laravel's
