--- app/Validation/CustomValidator.php
@@ -1,12 +1,0 @@
-<?php
-
-
-namespace App\Validation;
-
-
-interface CustomValidator
-{
-    public function passes($attribute, $value);
-
-    public function message($attribute);
-}
